extracts EmbeddedOptions, PredictButton and ResolveBy into their own components
